refactor: moved compile in compiler/
